feat(api): v3.6 PR10 — logging schema 확장 + miss_qps token bucket + timeout 정합
코덱스 PR8d review 의 잔여 P2 (FETCH_TIMEOUT_SEC 운영 근거 / 문서 SLA 불일치) +
Nit (rate_limited route 미문서화) 닫음. v3.5 step 3 §3.2 spec 의 logging schema
전체 적용.

P2 fix (timeout 정합):
- FETCH_TIMEOUT_SEC: 5.0 → 1.5 (v3.5 step 2 §291 spec).
- saatchi_detail_enricher 의 fetch_saatchi_html / fetch_and_parse_saatchi_detail
  의 timeout 시그니처 int → float (urllib.urlopen 도 float 허용).
- get_artwork_year 의 default fetch_fn 이 int 변환 없이 1.5 그대로 전달.
- saatchi 정상 응답 p95 600ms 안에서 충분 (1.5s margin).

PR8d deferred (miss_qps token bucket):
- FetchGate 에 token bucket 추가: capacity=3 (cold-start burst), refill=0.5/s
  (sustain target — v3.5 step 3 §3.2.3 의 miss_qps < 0.5 qps).
- _refill_tokens_locked: now - last_refill 차이만큼 token 보충 (capped capacity).
- try_acquire 평가 순서: cool_down → concurrent → burst → qps → inflight.
- stats() 에 tokens_available + inflight 노출.

Nit fix (route 문서):
- docs/v3_5_step3 line 132 의 route enum 에 'rate_limited' 추가.
- docs/v3_5_step3 §3.2.2 의 fallback_rate 정의에 rate_limited 포함 + 별도
  rate_limited_rate metric 추가 (목표 < 1%, 초과 시 burst / saatchi 진단).
- docs/v3_5_step2 §6.3 의 route enum 도 step3 와 정합되게 9 값 명시.

Logging schema 확장 (v3.5 step 3 §3.2 line 124-146 의 15 fields full):
- match_profile_source (profile.source — saatchi / artsy / null)
- slug_in_warm_set (predictor.is_warm_artist 결과 — cohort 와 별개로 관측)
- external_collector_source (sources_used 첫 source — 'none' fallback)
- artwork_id, artwork_url (request 그대로)
- enrichment_latency_ms (year resolve 단계 측정)
- predict_total_latency_ms (기존 total_ms 와 병기 — backward compat)
- model_variant (predictor.variant)
- artifact_version, warm_artist_slugs_version, rollout_rule_version,
  server_instance: env var 주입 (deploy pipeline 책임), 미설정 시 'unknown'
- cache_epoch: server cold-start 시점 (UTC ISO short)

테스트 (146 / 146 통과):
- token bucket 4 cases (block_after_capacity / refill / order / spec_values)
- timeout wire-up 1 case (FETCH_TIMEOUT_SEC=1.5 검증)
- logging schema 3 cases (full schema / artsy match / external_collector)
- 기존 cohort/cache test 의 token bucket 충돌은 fixture monkeypatch 로 우회
  (이 파일들은 concurrent/burst/cool-down/inflight/cache 검증 전용)

Lint: 변경 파일 clean. 기존 web_searcher / shap_explainer 위반은 PR10 범위 외.

코덱스 PR8 사이클 종결 조건 (P2 + Nit) 모두 closed. 다음 PR9 batch 진입 준비.

+++ scripts/saatchi_detail_enricher.py
@@ -195,7 +195,7 @@ def fetch_saatchi_html(
     url: str,
     *,
     user_agent: str = DEFAULT_UA,
-    timeout: int = DEFAULT_TIMEOUT,
+    timeout: float = DEFAULT_TIMEOUT,
 ) -> tuple[str, FetchStatus]:
     """URL 에서 raw HTML fetch + status 판정.
 
@@ -242,7 +242,7 @@ def fetch_and_parse_saatchi_detail(
     *,
     price_krw: float | int | None = None,
     user_agent: str = DEFAULT_UA,
-    timeout: int = DEFAULT_TIMEOUT,
+    timeout: float = DEFAULT_TIMEOUT,
 ) -> EnrichmentResult:
     """Fetch + parse wrapper. fetch 실패 시 빈 result 반환."""
     html, status = fetch_saatchi_html(url, user_agent=user_agent, timeout=timeout)
